+++ chaosk8s/actions.py
@@ -4,11 +4,13 @@ import json
 import os.path
 from typing import Union
 
-from kubernetes import client, config
-import yaml
-
 from chaoslib.exceptions import FailedProbe
 from chaoslib.types import MicroservicesStatus
+from kubernetes import client
+import yaml
+
+from chaosk8s import create_k8s_api_client
+
 
 __all__ = ["start_microservice", "kill_microservice",
            "remove_service_endpoint"]
@@ -19,7 +21,7 @@ def start_microservice(spec_path: str, ns: str = "default"):
     Start a microservice described by the deployment config, which must be the
     path to the JSON or YAML representation of the deployment.
     """
-    config.load_kube_config()
+    api = create_k8s_api_client()
 
     with open(spec_path) as f:
         p, ext = os.path.splitext(spec_path)
@@ -31,7 +33,7 @@ def start_microservice(spec_path: str, ns: str = "default"):
             raise FailedProbe(
                 "cannot process {path}".format(path=spec_path))
 
-    v1 = client.AppsV1beta1Api()
+    v1 = client.AppsV1beta1Api(api)
     resp = v1.create_namespaced_deployment(ns, body=deployment)
     return resp
 
@@ -46,31 +48,31 @@ def kill_microservice(name: str, ns: str = "default"):
     To work, the deployment must have a `service` label matching the
     `name` of the microservice.
     """
-    config.load_kube_config()
+    api = create_k8s_api_client()
 
-    v1 = client.AppsV1beta1Api()
+    v1 = client.AppsV1beta1Api(api)
     ret = v1.list_namespaced_deployment(
         ns, label_selector="service={name}".format(name=name))
 
-    body = client.V1DeleteOptions()
+    body = client.V1DeleteOptions(api)
     for d in ret.items:
         res = v1.delete_namespaced_deployment(
             d.metadata.name, ns, body)
 
-    v1 = client.ExtensionsV1beta1Api()
+    v1 = client.ExtensionsV1beta1Api(client)
     ret = v1.list_namespaced_replica_set(
         ns, label_selector="service={name}".format(name=name))
 
-    body = client.V1DeleteOptions()
+    body = client.V1DeleteOptions(api)
     for r in ret.items:
         res = v1.delete_namespaced_replica_set(
             r.metadata.name, ns, body)
 
-    v1 = client.CoreV1Api()
+    v1 = client.CoreV1Api(api)
     ret = v1.list_namespaced_pod(
         ns, label_selector="service={name}".format(name=name))
 
-    body = client.V1DeleteOptions()
+    body = client.V1DeleteOptions(api)
     for p in ret.items:
         res = v1.delete_namespaced_pod(
             p.metadata.name, ns, body)
@@ -80,7 +82,7 @@ def remove_service_endpoint(name: str, ns: str = "default"):
     """
     Remove the service endpoint that sits in front of microservices (pods).
     """
-    config.load_kube_config()
+    api = create_k8s_api_client()
 
-    v1 = client.CoreV1Api()
+    v1 = client.CoreV1Api(api)
     v1.delete_namespaced_service(name, namespace=ns)