+++ chaosk8s/probes.py
@@ -3,11 +3,13 @@ import json
 import os.path
 from typing import Union
 
-from kubernetes import client, config
-import yaml
-
 from chaoslib.exceptions import FailedProbe
 from chaoslib.types import MicroservicesStatus
+from kubernetes import client
+import yaml
+
+from chaosk8s import create_k8s_api_client
+
 
 __all__ = ["all_microservices_healthy", "microservice_available_and_healthy",
            "microservice_is_not_available", "service_endpoint_is_initialized"]
@@ -20,11 +22,11 @@ def all_microservices_healthy(ns: str = "default") -> MicroservicesStatus:
     Raises :exc:`chaoslib.exceptions.FailedProbe` when the state is not
     as expected.
     """
-    config.load_kube_config()
+    api = create_k8s_api_client()
     not_ready = []
     failed = []
 
-    v1 = client.CoreV1Api()
+    v1 = client.CoreV1Api(api)
     ret = v1.list_namespaced_pod(namespace=ns)
     for p in ret.items:
         phase = p.status.phase
@@ -47,9 +49,9 @@ def microservice_available_and_healthy(
     Raises :exc:`chaoslib.exceptions.FailedProbe` when the state is not
     as expected.
     """
-    config.load_kube_config()
+    api = create_k8s_api_client()
 
-    v1 = client.AppsV1beta1Api()
+    v1 = client.AppsV1beta1Api(api)
     ret = v1.list_namespaced_deployment(
         ns, label_selector="service={name}".format(name=name))
 
@@ -71,9 +73,9 @@ def microservice_is_not_available(name: str, ns: str = "default") -> bool:
     Raises :exc:`chaoslib.exceptions.FailedProbe` when the state is not
     as expected.
     """
-    config.load_kube_config()
+    api = create_k8s_api_client()
 
-    v1 = client.AppsV1beta1Api()
+    v1 = client.AppsV1beta1Api(api)
     ret = v1.list_namespaced_deployment(
         ns, label_selector="service={name}".format(name=name))
 
@@ -87,9 +89,9 @@ def service_endpoint_is_initialized(name: str, ns: str= "default"):
     Lookup a service endpoint by its name and raises :exc:`FailedProbe` when
     the service was not found or not initialized.
     """
-    config.load_kube_config()
+    api = create_k8s_api_client()
 
-    v1 = client.CoreV1Api()
+    v1 = client.CoreV1Api(api)
     ret = v1.list_namespaced_service(
         ns, label_selector="service={name}".format(name=name))